+++ iscc_schema/__init__.py
@@ -3,3 +3,4 @@ from iscc_schema.schema import IsccMeta, Signature
 from iscc_schema.seed_isbn import ISBN
 from iscc_schema.seed_isrc import ISRC
 from iscc_schema.service_tdm import TDM
+from iscc_schema.recovery import recover_context